requirements.txt
dash==2.6.2
PyYAML==6.0
dash-bootstrap-components==1.2.1
dash-iconify==0.1.2